+++ packages/mui/src/FieldTemplate/FieldTemplate.tsx
@@ -1,5 +1,5 @@
-import FormControl from '@mui/material/FormControl';
-import Typography from '@mui/material/Typography';
+import FormControl, { FormControlProps } from '@mui/material/FormControl';
+import Typography, { TypographyProps } from '@mui/material/Typography';
 import {
   FieldTemplateProps,
   FormContextType,
@@ -8,6 +8,7 @@ import {
   getTemplate,
   getUiOptions,
 } from '@rjsf/utils';
+import { getMuiProps } from '../util';
 
 /** The `FieldTemplate` component is the template used by `SchemaField` to render any field. It renders the field
  * content, (label, description, children, errors and help) inside of a `WrapIfAdditional` component.
@@ -55,6 +56,13 @@ export default function FieldTemplate<
 
   const isCheckbox = uiOptions.widget === 'checkbox';
 
+  const muiProps = getMuiProps<T, S, F>({
+    uiSchema,
+    formContext: registry.formContext,
+    options: uiOptions,
+  });
+  const { slotProps: muiSlotProps, ...otherMuiProps } = muiProps;
+
   return (
     <WrapIfAdditionalTemplate
       classNames={classNames}
@@ -73,10 +81,17 @@ export default function FieldTemplate<
       uiSchema={uiSchema}
       registry={registry}
     >
-      <FormControl fullWidth={true} error={rawErrors.length ? true : false} required={required}>
+      <FormControl
+        fullWidth={true}
+        error={rawErrors.length ? true : false}
+        required={required}
+        {...(muiSlotProps?.formControl as FormControlProps)}
+        sx={otherMuiProps.sx}
+        className={otherMuiProps.className}
+      >
         {children}
         {displayLabel && !isCheckbox && rawDescription ? (
-          <Typography variant='caption' color='textSecondary'>
+          <Typography variant='caption' color='textSecondary' {...(muiSlotProps?.typography as TypographyProps)}>
             {description}
           </Typography>
         ) : null}