+++ packages/mui/src/IconButton/IconButton.tsx
@@ -5,6 +5,7 @@ import CopyIcon from '@mui/icons-material/ContentCopy';
 import RemoveIcon from '@mui/icons-material/Remove';
 import ClearIcon from '@mui/icons-material/Clear';
 import { FormContextType, IconButtonProps, RJSFSchema, StrictRJSFSchema, TranslatableString } from '@rjsf/utils';
+import { getMuiProps } from '../util';
 
 export default function MuiIconButton<
   T = any,
@@ -12,8 +13,19 @@ export default function MuiIconButton<
   F extends FormContextType = any,
 >(props: IconButtonProps<T, S, F>) {
   const { icon, color, uiSchema, registry, ...otherProps } = props;
+
+  const muiProps = getMuiProps<T, S, F>({
+    uiSchema,
+    formContext: registry.formContext,
+  });
+
   return (
-    <IconButton {...otherProps} size='small' color={color as MuiIconButtonProps['color']}>
+    <IconButton
+      {...(muiProps as MuiIconButtonProps)}
+      {...otherProps}
+      size='small'
+      color={color as MuiIconButtonProps['color']}
+    >
       {icon}
     </IconButton>
   );